requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
apischema==0.17.5
attrs==24.3.0
base58==2.1.1
based58==0.1.1
bitarray==3.0.0
cachetools==4.2.4
certifi==2024.12.14
charset-normalizer==3.4.1
ckzg==2.0.1
construct==2.10.68
construct-typing==0.5.6
cytoolz==1.0.1
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.1.0
eth_abi==5.2.0
frozenlist==1.5.0
h11==0.14.0
hexbytes==1.3.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
jsonalias==0.1.1
jsonrpcclient==4.0.3
jsonrpcserver==5.0.9
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
multidict==6.1.0
OSlash==0.6.3
parsimonious==0.10.0
propcache==0.2.1
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rlp==4.0.1
rpds-py==0.22.3
sniffio==1.3.1
solana-sdk==0.25.6
solders==0.2.0
toolz==1.0.0
types-cachetools==4.2.10
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
web3==7.7.0
websockets==10.4
yarl==1.18.3
